simulation.py: make a tiny argument parser.
My test laptop has a high-DPI screen, so it's useful to be able to
make the whole simulation window bigger, by making `LED_SIZE`
configurable on the command line.

While I'm at it, I've also added an option to specify the path of
simulation.so, because that kind of thing always seems to come in
handy sooner or later.