Use an X-macro for the list of patterns.
The list of pattern names is now defined just once, in common_led.h,
in the form of a macro that takes another macro as a parameter. So you
can iterate over the list of pattern names doing a thing for each one,
basically by defining a macro to do your thing, and invoking
ALL_PATTERNS(that-macro).

This allows autogenerating a list of instances of all the classes, and
a list of if statements that checks a string against all the class
names, and even the list of available pattern names in the hardcoded
JSON in mqtt.cpp. But for the last of those, I had to also give
ALL_PATTERNS a second argument to interleave between the invocations
of the parameter macro, because otherwise, you wouldn't be able to
avoid having a trailing comma at the end of the JSON list, which isn't
legal JSON.